move formatUrlWithWorkspaceId from plugin workspace to core
