+++ src/core/public/chrome/constants.ts
@@ -31,4 +31,3 @@
 export const OPENSEARCH_DASHBOARDS_ASK_OPENSEARCH_LINK = 'https://forum.opensearch.org/';
 export const GITHUB_CREATE_ISSUE_LINK =
   'https://github.com/opensearch-project/OpenSearch-Dashboards/issues/new/choose';
-export const WORKSPACE_OVERVIEW_APP_ID = 'workspace_overview';
