+++ src/plugins/workspace/public/components/workspace_updater/workspace_updater.tsx
@@ -20,7 +20,7 @@ import { useOpenSearchDashboards } from '../../../../opensearch_dashboards_react
 import { WorkspaceForm, WorkspaceFormData } from '../workspace_creator/workspace_form';
 import { WORKSPACE_OVERVIEW_APP_ID, WORKSPACE_OP_TYPE_UPDATE } from '../../../common/constants';
 import { DeleteWorkspaceModal } from '../delete_workspace_modal';
-import { formatUrlWithWorkspaceId } from '../../utils';
+import { formatUrlWithWorkspaceId } from '../../../../../core/public/utils';
 import { WorkspaceClient } from '../../workspace_client';
 
 export const WorkspaceUpdater = () => {
